# app/api/api.yaml
openapi: 3.0.0
info:
  title: Koderpedia ERP API
  version: 1.0.0
servers:
  - url: 'http://localhost:1323'
    description: Local development server
...
paths:
...
  /api/v1/auth/api-contracts:
    get:
      tags:
        - Auth Access
      operationId: ListAuthApiContracts
      summary: List API contracts
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: API contracts listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuthApiContractsResponse'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    post:
      tags:
        - Auth Access
      operationId: CreateAuthApiContract
      summary: Create API contract
      description: Requires Authorization header and a user API contract grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthApiContractRequest'
      responses:
        '201':
          description: API contract created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthApiContract'
        '400':
          description: Bad request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
  '/api/v1/auth/api-contracts/{id}':
    parameters:
      - name: id
        in: path
        required: true
        description: API contract id
        schema:
          type: string
          example: CryptoEncryptText
    get:
      tags:
        - Auth Access
      operationId: GetAuthApiContract
      summary: Get API contract
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: API contract returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthApiContract'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    put:
      tags:
        - Auth Access
      operationId: UpdateAuthApiContract
      summary: Update API contract
      description: Requires Authorization header and a user API contract grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuthApiContractRequest'
      responses:
        '200':
          description: API contract updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthApiContract'
        '400':
          description: Bad request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    delete:
      tags:
        - Auth Access
      operationId: DeleteAuthApiContract
      summary: Delete API contract
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: API contract deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAuthAccessResponse'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
...
  /api/v1/auth/roles:
    get:
      tags:
        - Auth Roles
      operationId: ListAuthRoles
      summary: List auth roles
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: Auth roles listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuthRolesResponse'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    post:
      tags:
        - Auth Roles
      operationId: CreateAuthRole
      summary: Create auth role
      description: Requires Authorization header and a user API contract grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthRoleRequest'
      responses:
        '201':
          description: Auth role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRole'
        '400':
          description: Bad request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
  '/api/v1/auth/roles/{id}':
    parameters:
      - name: id
        in: path
        required: true
        description: Auth role id
        schema:
          type: string
          example: 1f2f4181-b4b1-4812-9f57-a6a315654040
    get:
      tags:
        - Auth Roles
      operationId: GetAuthRole
      summary: Get auth role
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: Auth role returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRole'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    put:
      tags:
        - Auth Roles
      operationId: UpdateAuthRole
      summary: Update auth role
      description: Requires Authorization header and a user API contract grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuthRoleRequest'
      responses:
        '200':
          description: Auth role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRole'
        '400':
          description: Bad request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    delete:
      tags:
        - Auth Roles
      operationId: DeleteAuthRole
      summary: Delete auth role
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: Auth role deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAuthAccessResponse'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
  '/api/v1/auth/roles/{roleId}/api-contracts':
    parameters:
      - name: roleId
        in: path
        required: true
        description: Auth role id
        schema:
          type: string
          example: 1f2f4181-b4b1-4812-9f57-a6a315654040
    get:
      tags:
        - Auth Roles
      operationId: ListAuthRoleContractApis
      summary: List role API contracts
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: Role API contracts listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuthRoleContractApisResponse'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    post:
      tags:
        - Auth Roles
      operationId: CreateAuthRoleContractApi
      summary: Assign API contract to role
      description: Requires Authorization header and a user API contract grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthRoleContractApiRequest'
      responses:
        '201':
          description: Role API contract created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRoleContractApi'
        '400':
          description: Bad request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
  '/api/v1/auth/roles/{roleId}/api-contracts/{id}':
    parameters:
      - name: roleId
        in: path
        required: true
        description: Auth role id
        schema:
          type: string
          example: 1f2f4181-b4b1-4812-9f57-a6a315654040
      - name: id
        in: path
        required: true
        description: Role API contract mapping id
        schema:
          type: string
          example: 31267553-e12a-497d-bc4a-767b642f4dd6
    delete:
      tags:
        - Auth Roles
      operationId: DeleteAuthRoleContractApi
      summary: Remove API contract from role
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: Role API contract deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAuthAccessResponse'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
...
components:
  schemas:
...
    ErrorBadRequest:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message describing the bad request
          example: Invalid input data
    ErrorUnauthorized:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message describing the unauthorized access
          example: 'Unauthorized, invalid OTP code'
    ErrorForbidden:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message describing the forbidden access
          example: 'Forbidden, you do not have permission to access this resource'
...
    AuthApiContract:
      type: object
      required:
        - id
        - endpointPath
        - endpointMethod
        - description
        - createdAtUtc0
        - updatedAtUtc0
        - isActive
      properties:
        id:
          type: string
          description: OpenAPI operationId for the API contract
          example: CryptoEncryptText
        endpointPath:
          type: string
          description: Echo route path for the API contract
          example: /api/v1/crypto/encrypt
        endpointMethod:
          type: string
          description: HTTP method for the API contract
          example: post
        description:
          type: string
          description: Description of the API contract purpose
          example: Encrypt plain text using configured crypto service
        createdAtUtc0:
          type: integer
          format: int64
          description: Created at in UTC+0 Unix millisecond time
          example: 1798790400000
        updatedAtUtc0:
          type: integer
          format: int64
          description: Updated at in UTC+0 Unix millisecond time
          example: 1798790400000
        isActive:
          type: boolean
          description: Whether the API contract is active
          example: true
    CreateAuthApiContractRequest:
      type: object
      required:
        - id
        - endpointPath
        - endpointMethod
        - description
      properties:
        id:
          type: string
          description: OpenAPI operationId for the API contract
          example: CryptoEncryptText
        endpointPath:
          type: string
          description: Echo route path for the API contract
          example: /api/v1/crypto/encrypt
        endpointMethod:
          type: string
          description: HTTP method for the API contract
          example: post
        description:
          type: string
          description: Description of the API contract purpose
          example: Encrypt plain text using configured crypto service
    UpdateAuthApiContractRequest:
      type: object
      required:
        - endpointPath
        - endpointMethod
        - description
        - isActive
      properties:
        endpointPath:
          type: string
          description: Echo route path for the API contract
          example: /api/v1/crypto/encrypt
        endpointMethod:
          type: string
          description: HTTP method for the API contract
          example: post
        description:
          type: string
          description: Description of the API contract purpose
          example: Encrypt plain text using configured crypto service
        isActive:
          type: boolean
          description: Whether the API contract is active
          example: true
    ListAuthApiContractsResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          description: Active API contracts
          items:
            $ref: '#/components/schemas/AuthApiContract'
          example:
            - id: CryptoEncryptText
              endpointPath: /api/v1/crypto/encrypt
              endpointMethod: post
              description: Encrypt plain text using configured crypto service
              createdAtUtc0: 1798790400000
              updatedAtUtc0: 1798790400000
              isActive: true
...
    DeleteAuthAccessResponse:
      type: object
      required:
        - isSuccess
      properties:
        isSuccess:
          type: boolean
          description: Whether the delete operation succeeded
          example: true
    AuthRole:
      type: object
      required:
        - id
        - roleName
        - description
        - ownerId
        - createdAtUtc0
        - updatedAtUtc0
        - createdBy
        - isActive
      properties:
        id:
          type: string
          description: Auth role id
          example: 1f2f4181-b4b1-4812-9f57-a6a315654040
        roleName:
          type: string
          description: Name of the role
          example: admin
        description:
          type: string
          description: Description of the role purpose
          example: Administrator role with management access
        ownerId:
          type: string
          description: Auth user id that owns the role
          example: 358cbaad-316e-4539-9949-2636cdbd7e89
        createdAtUtc0:
          type: integer
          format: int64
          description: Created at in UTC+0 Unix millisecond time
          example: 1798790400000
        updatedAtUtc0:
          type: integer
          format: int64
          description: Updated at in UTC+0 Unix millisecond time
          example: 1798790400000
        createdBy:
          type: string
          description: Creator identifier
          example: admin-user
        isActive:
          type: boolean
          description: Whether the role is active
          example: true
    CreateAuthRoleRequest:
      type: object
      required:
        - roleName
        - description
        - ownerId
        - createdBy
      properties:
        roleName:
          type: string
          description: Name of the role
          example: admin
        description:
          type: string
          description: Description of the role purpose
          example: Administrator role with management access
        ownerId:
          type: string
          description: Auth user id that owns the role
          example: 358cbaad-316e-4539-9949-2636cdbd7e89
        createdBy:
          type: string
          description: Creator identifier
          example: admin-user
    UpdateAuthRoleRequest:
      type: object
      required:
        - roleName
        - description
        - ownerId
        - createdBy
        - isActive
      properties:
        roleName:
          type: string
          description: Name of the role
          example: admin
        description:
          type: string
          description: Description of the role purpose
          example: Administrator role with management access
        ownerId:
          type: string
          description: Auth user id that owns the role
          example: 358cbaad-316e-4539-9949-2636cdbd7e89
        createdBy:
          type: string
          description: Creator identifier
          example: admin-user
        isActive:
          type: boolean
          description: Whether the role is active
          example: true
    ListAuthRolesResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          description: Active auth roles
          items:
            $ref: '#/components/schemas/AuthRole'
          example:
            - id: 1f2f4181-b4b1-4812-9f57-a6a315654040
              roleName: admin
              description: Administrator role with management access
              ownerId: 358cbaad-316e-4539-9949-2636cdbd7e89
              createdAtUtc0: 1798790400000
              updatedAtUtc0: 1798790400000
              createdBy: admin-user
              isActive: true
    AuthRoleContractApi:
      type: object
      required:
        - id
        - roleId
        - authApiContractId
        - createdAtUtc0
        - updatedAtUtc0
        - createdBy
        - isActive
      properties:
        id:
          type: string
          description: Role API contract mapping id
          example: 31267553-e12a-497d-bc4a-767b642f4dd6
        roleId:
          type: string
          description: Auth role id
          example: 1f2f4181-b4b1-4812-9f57-a6a315654040
        authApiContractId:
          type: string
          description: API contract id assigned to the role
          example: CryptoEncryptText
        createdAtUtc0:
          type: integer
          format: int64
          description: Created at in UTC+0 Unix millisecond time
          example: 1798790400000
        updatedAtUtc0:
          type: integer
          format: int64
          description: Updated at in UTC+0 Unix millisecond time
          example: 1798790400000
        createdBy:
          type: string
          description: Creator identifier
          example: admin-user
        isActive:
          type: boolean
          description: Whether the role API contract mapping is active
          example: true
    CreateAuthRoleContractApiRequest:
      type: object
      required:
        - authApiContractId
        - createdBy
      properties:
        authApiContractId:
          type: string
          description: API contract id assigned to the role
          example: CryptoEncryptText
        createdBy:
          type: string
          description: Creator identifier
          example: admin-user
    ListAuthRoleContractApisResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          description: Active role API contract mappings
          items:
            $ref: '#/components/schemas/AuthRoleContractApi'
          example:
            - id: 31267553-e12a-497d-bc4a-767b642f4dd6
              roleId: 1f2f4181-b4b1-4812-9f57-a6a315654040
              authApiContractId: CryptoEncryptText
              createdAtUtc0: 1798790400000
              updatedAtUtc0: 1798790400000
              createdBy: admin-user
              isActive: true
...